refactor: Branch from component
