+++ younghyun/src/pages/auth/loginPresenter.jsx
@@ -1,6 +1,8 @@
 import styled from "styled-components";
 
-const LoginContainer = styled.div``;
+const LoginContainer = styled.div`
+    border: 1px solid black;
+`;
 
 function LoginPresenter() {
     return (
